modify data length

+++ detail/common/spin_detail.go
@@ -7,8 +7,8 @@ import (
 type Game struct {
 	GameId      string `json:"game_id" orm:"pk"`
 	GameName    string `json:"game_name"`
-	Rows        uint   `json:"rows"`
-	Columns     uint   `json:"columns"`
+	Rows        uint8   `json:"rows"`
+	Columns     uint8   `json:"columns"`
 	ColumnsInfo string `json:"columns_info"` // json数组 [int]
 
 	CreatedAt time.Time `json:"-" orm:"auto_now_add;type(datetime)"`
@@ -46,14 +46,14 @@ func (u *Round) TableName() string {
 type SpinNew struct {
 	Id         uint64 `json:"-" orm:"pk;auto"`
 	Round      *Round `json:"-" orm:"rel(fk)"`
-	FreeGameId uint   `json:"free_game_id"` // 主游戏填0 freespin填具体值第几次freespin
-	RespinId   uint   `json:"respin_id"`    //非resipin情况下填0， respin情况下填具体第几次respin，从1开始计
-	SpinType   uint   `json:"spin_type"`    //子游戏类型  0 主游戏， 1 freespin ，2 主游戏中的respin 3 freespin中现的espin
+	FreeGameId uint8   `json:"free_game_id"` // 主游戏填0 freespin填具体值第几次freespin
+	RespinId   uint8   `json:"respin_id"`    //非resipin情况下填0， respin情况下填具体第几次respin，从1开始计
+	SpinType   uint8   `json:"spin_type"`    //子游戏类型  0 主游戏， 1 freespin ，2 主游戏中的respin 3 freespin中现的espin
 
 	SpinReward       uint      `json:"spin_reward"`                   //此次spin赢钱
 	SpinBet          uint      `json:"spin_bet"`                      //此次下注金额 主游戏有效，freespin时无效,对于美人鱼 辣椒 和红唇，该值和TotalReward相同
 	BetTime          time.Time `json:"bet_time" orm:"type(datetime)"` //下注时间 XXXX/XX/XX XX:XX:XX
-	RewardLineNumber uint      `json:"reward_line_number"`            //赢钱线数
+	RewardLineNumber uint8      `json:"reward_line_number"`            //赢钱线数
 
 	Items         string           `json:"items" orm:"size(10000)"`            //json数组 [[int]]                       //旋转结果，一个子数组字代表一列
 	RewardDetails []*GenericReward `json:"reward_details" orm:"reverse(many)"` //包含中scatter bonus jackpot 线等各种中奖
@@ -74,15 +74,15 @@ func (u *SpinNew) TableUnique() [][]string {
 
 type GenericReward struct {
 	Id int64 `json:"-" orm:"pk;auto"`
-									 //Round         *Round   `json:"-" orm:"rel(fk)"`
+									  //Round         *Round   `json:"-" orm:"rel(fk)"`
 	Spin          *SpinNew `json:"-" orm:"rel(fk)"`
 	LineId        uint     `json:"line_id"`                          //中奖线id  如果是bonus 或者 scatter中奖，就把该值设为 bonus 和 scatter图标
-	RewardType    uint     `json:"reward_type"`                      // 0 item中奖 1 scatter  2 bonus 3 jackpot
+	RewardType    uint8     `json:"reward_type"`                      // 0 item中奖 1 scatter  2 bonus 3 jackpot
 	RewardItems   string   `json:"reward_items"`                     //json数组 [int] 中奖图标 对于美人鱼，只有一个图标；红唇会有多个图标
 	BetMultiple   uint     `json:"bet_multiple"`                     //下注倍数
 	Reward        uint     `json:"reward"`                           //此线赢钱
 	Multiple      uint     `json:"multiple"`                         // 根据玩法，玩法中如果有乘倍则显示倍数，如果没有则显示1（现阶段只有财神机器有乘倍）
-	ItemNumber    uint     `json:"item_number"`                      //几连线
+	ItemNumber    uint8     `json:"item_number"`                      //几连线
 	ItemPositions string   `json:"item_positions" orm:"size(10000)"` // json数组 [[column,row]]
 
 	CreatedAt time.Time `json:"-" orm:"auto_now_add;type(datetime)"`
@@ -115,9 +115,10 @@ func (u *GameStat) TableName() string {
 	return "t_game_stat_reward"
 }
 
-
+/*
 func (u *GameStat) TableUnique() [][]string {
 	return [][]string{
 		[]string{"Round"},
 	}
 }
+*/
